pyproject.toml
[tool.poetry]
name = "llm_test_package"
version = "0.1.0"
description = ""
authors = ["Nico Manthey <your.email@example.com>"]

[tool.poetry.dependencies]
python = "3.11.9"
gradio = "*"
requests = "*"
huggingface_hub = "*"
streamlit = "^1.36.0"
pandas = "^2.2.2"
google-auth = "^2.30.0"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.134.0"
gspread = "^6.1.2"
watchdog = "^4.0.1"

[tool.poetry.dev-dependencies]
pytest = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"